{"schema":"decodex/commit/1","summary":"Split consolidation proposal types","authority":"manual"}
